Remove legacy jira client. Use Jira native models and move ours to models.ts. Rename stream to task_pull_requests.
