+++ destinations/airbyte-faros-destination/src/converters/jira/task_pull_requests.ts
@@ -5,7 +5,7 @@ import {DestinationModel, DestinationRecord, StreamContext} from '../converter';
 import {JiraConverter} from './common';
 
 // Required as dependency by Issues converter
-export class PullRequests extends JiraConverter {
+export class TaskPullRequests extends JiraConverter {
   readonly destinationModels: ReadonlyArray<DestinationModel> = [
     'tms_TaskPullRequestAssociation',
   ];
